Rename namespace to CMS and change folder name to CMS
