+++ Phoenix/Client/Include/Client/InputMap.hpp
@@ -32,7 +32,7 @@
 #include <Client/Events/IEventListener.hpp>
 
 #include <Common/Singleton.hpp>
-#include <Common/Mods/ModManager.hpp>
+#include <Common/CMS/ModManager.hpp>
 
 #include <functional>
 #include <string>
