+++ Phoenix/Common/Include/Common/CMS/CMakeLists.txt
@@ -1,5 +1,5 @@
 set(currentDir ${CMAKE_CURRENT_LIST_DIR})
-set(modHeaders
+set(cmsHeaders
 	${currentDir}/Mod.hpp
 	${currentDir}/ModAPI.hpp
 	${currentDir}/ModManager.hpp
