+++ Phoenix/Common/Source/CMS/CMakeLists.txt
@@ -1,5 +1,5 @@
 set(currentDir ${CMAKE_CURRENT_LIST_DIR})
-set(modSources
+set(cmsSources
 	${currentDir}/Mod.cpp
 	${currentDir}/ModAPI.cpp
 	${currentDir}/ModManager.cpp
